requirements.txt
aiohttp
aiohttp-cors
pytest
janus
rainbow_logging_handler
pyyaml
git+https://github.com/umd-lhcb/rpi.burnin
git+https://github.com/umd-lhcb/labSNMP
